audio module decomposed; audio streaming rewritten; work in progress...
