+++ src/gui/gui.cpp
@@ -15,13 +15,13 @@
 #include "../render/shader_description.h"
 #include "../render/shader_manager.h"
 #include "../script/script.h"
+#include "../audio/audio.h"
 #include "../image.h"
 #include "../mesh.h"
 #include "../skeletal_model.h"
 #include "../entity.h"
 #include "../character_controller.h"
 #include "../engine.h"
-#include "../audio.h"
 #include "../engine_string.h"
 #include "../world.h"
 #include "../inventory.h"
