+++ src/state_control/state_control_bat.cpp
@@ -10,7 +10,6 @@
 #include "../script/script.h"
 #include "../vt/tr_versions.h"
 #include "../engine.h"
-#include "../audio.h"
 #include "../controls.h"
 #include "../room.h"
 #include "../world.h"
