+++ src/trigger.cpp
@@ -15,12 +15,12 @@ extern "C" {
 #include "core/vmath.h"
 #include "script/script.h"
 #include "render/camera.h"
+#include "audio/audio.h"
 #include "engine.h"
 #include "room.h"
 #include "trigger.h"
 #include "gameflow.h"
 #include "game.h"
-#include "audio.h"
 #include "skeletal_model.h"
 #include "entity.h"
 #include "character_controller.h"
