+++ src/world.cpp
@@ -23,12 +23,12 @@ extern "C" {
 #include "render/frustum.h"
 #include "render/render.h"
 #include "render/bordered_texture_atlas.h"
+#include "audio/audio.h"
 #include "script/script.h"
 #include "state_control/state_control.h"
 #include "physics/physics.h"
 #include "gui/gui.h"
 #include "vt/vt_level.h"
-#include "audio.h"
 #include "room.h"
 #include "world.h"
 #include "mesh.h"
